+++ apps/website/src/App.tsx
@@ -1,9 +1,8 @@
-import { getTheme, vars } from '@sledge/theme';
+import { getTheme } from '@sledge/theme';
 import { MetaProvider } from '@solidjs/meta';
 import { Route, Router } from '@solidjs/router';
 import { onMount, Suspense, type Component } from 'solid-js';
-import Home from '~/routes';
-import Telling from '~/routes/telling';
+import { Start } from '~/routes/index';
 import { globalStore } from '~/store/GlobalStore';
 
 const App: Component = () => {
@@ -17,22 +16,14 @@ const App: Component = () => {
         <MetaProvider>
           <title>Sledge</title>
           <Suspense>
-            <main
-              class={getTheme(globalStore.theme)}
-              style={{
-                background: vars.color.surface,
-              }}
-            >
-              {props.children}
-            </main>
+            <main class={getTheme(globalStore.theme)}>{props.children}</main>
           </Suspense>
 
           {/* <p class={sledgeLogo}>sledge.</p> */}
         </MetaProvider>
       )}
     >
-      <Route path='/' component={Home} />
-      <Route path='/telling' component={Telling} />
+      <Route path='/' component={Start} />
     </Router>
   );
 };
