+++ apps/website/src/routes/README.md
@@ -1,0 +1,6 @@
+# Telling Sections
+
+1. Hard Edged
+2. Chaos Engine
+3. Open Core
+4. Code Your Canvas
