+++ apps/website/src/store/GlobalStore.ts
@@ -3,6 +3,8 @@ import { createStore } from 'solid-js/store';
 
 export const [globalStore, setGlobalStore] = createStore<{
   theme: Theme;
+  currentTellingPage: number;
 }>({
   theme: (localStorage.getItem('theme') as Theme) || 'dark-gy-flip',
+  currentTellingPage: 0,
 });
